milestone3/yizhen/src/model-training/pyproject.toml
[project]
name = "dnd-model-trainer"
version = "0.1.0"
description = "DnD Model Training package for Vertex AI"
readme = "README.md"
requires-python = ">=3.10,<3.13"
authors = [
    { name = "Yizhen Wang", email = "your_email@example.com" }
]
license = { text = "MIT" }

dependencies = [
    # Core ML
    "tensorflow==2.15.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",

    # Cloud integration
    "google-cloud-storage>=2.16.0",
    "google-cloud-aiplatform==1.120.0",
    "google-auth>=2.25.0",

    # Utilities
    "pillow>=10.0.0",
    "tqdm>=4.65.0",
    "requests>=2.31.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
dependencies = [
    # Core ML
    "tensorflow==2.15.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",

    # Cloud integration
    "google-cloud-storage>=2.16.0",
    "google-cloud-aiplatform==1.120.0",
    "google-auth>=2.25.0",

    # Utilities
    "pillow>=10.0.0",
    "tqdm>=4.65.0",
    "requests>=2.31.0",
]
